chore:주문하기(진행중 - customerOrder DB에 값 넣기)

+++ src/main/java/com/jujeob/entity/OrderItem.java
@@ -25,7 +25,7 @@ public class OrderItem {
     @JoinColumn(name = "productno", referencedColumnName = "productno") // 수정
     private Product product; // 수정*/
 
-    private Long productNo;
+    private Integer productNo;
     private int quantity;
     private double price;
